refactor: type collaboration artifact placement
